final code cleanup and bugs fix
