+++ src/app/u/moje-knihy/[id]/page.tsx
@@ -1,6 +1,5 @@
 "use server";
 
-import { getPouzivateloveKnihy } from "@/lib/actions";
 import prisma from "@/lib/prisma";
 import { Typography } from "@mui/material";
 
@@ -25,20 +24,9 @@ export default async function Page({ params }
     return (
         pouzivatelovaKniha ? <>
             <Typography variant="h1">{pouzivatelovaKniha.kniha.nazov}</Typography>
-            <Typography variant="h5">Stav knihy {pouzivatelovaKniha.stav === "chcemPrecitat" ? "Na prečítanie" : "Prečítané"}</Typography>
-            <Typography variant="h5">Poznámka: {pouzivatelovaKniha.poznamka}</Typography>
+            <Typography variant="h5">Stav knihy: {pouzivatelovaKniha.stav === "chcemPrecitat" ? "na prečítanie" : "prečítané"}</Typography>
+            <Typography variant="h5">Poznámka: {pouzivatelovaKniha.poznamka || "Bez poznámky"}</Typography>
         </>
             : <Typography variant="h1">Používateľova kniha neexistuje</Typography>
     )
-
-    // return (
-    //     data ?
-    //         <Box component={"form"} action={updateKniha}>
-    //             <Typography variant="h1">{data?.nazov}</Typography>
-    //             <TextField name="pocet_stran" label="Počet strán" defaultValue={data?.pocet_stran} />
-    //             <Typography variant="body1">Počet strán: {data?.pocet_stran}</Typography>
-    //             <Typography variant="body1">Autor: {data?.autor.meno} {data?.autor.priezvisko}</Typography>
-    //             <Button type="submit" variant="contained">Aktualizovať knihu</Button>
-    //         </Box> : <Typography variant="h1">Kniha neexistuje</Typography>
-    // )
 }
